% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV format file into a data frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of a file.  Relative
or absolute paths may be used to locate the file in the string.}
}
\value{
This function returns a data frame constructed from the file contents.
}
\description{
The function determines if \code{filename} argument) exists and stops grecefully
if it does not.  If \code{filename} exists than the contents are returned as a
data frame.  If \code{filename} does not exist then nothing is returned, or assigned, but
no error or warning is generated.
}
\examples{

\dontrun{
#' fars_read(inputfile)
fars_read("another.csv")
}

}
