% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads a sequence of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An list of interger values (or values that will cast to integer) repersenting years.}
}
\value{
This function returns a list of tibbles constructed from the data contained in the file for
each year.
}
\description{
The function an argument \code{years} and builds a data fileneame out of it using \code{make_filename}
Uses \code{dplyr::mutate} and \code{dplyr::select}.  A list with the same number of elements as \code{years}
is returned with the contents of the respective file in each element.  If a file does not exist with the
name generated by \code{make_filename} then a warning is issued and that position in the returned list is
NULL.
}
\examples{
\dontrun{#' fars_read_years(c(2013,2015)
fars_read_years(2014)
}
}
