% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generates a table (tibble) summarizing the number of accidents by year and month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An list of interger values (or values that will cast to integer) repersenting years.}
}
\value{
This function returns an (n+1) x 12 tibble where n is the number of elements in the
argument \code{years}.  The additional column is the month number.  The value of each cell is
the count of accidents by month (row) and year (column).
}
\description{
The function uses the argument \code{years} and returns a summary of accidents for \code{years}.
}
\examples{
\dontrun{#' fars_summarize_years(c(2013,2015)
fars_summarize_years(2014)
}
}
