% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates a file name based on a year and standard pattern.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value repersenting the year.}
}
\value{
This function returns a text string file name constructed from the the integer
year embedded into a pattern.
}
\description{
The function takes the argument \code{year} and casts it to an integer before inserting it within
the pattern \code{accident<year>.csv.bz2} and returning the result.  If \code{year} cannot be cast
to an integer then 'NA' is placed in the file name for the integer year.
}
\examples{
yearfile <- make_filename(current_year)
yearfile2 <- make_filename(2010)
yearfile2 <- make_filename("2010")

}
